move ITS tool development to THAPBI-pycits repository
This is a complete reconstruction (with history) of the `santi_script`
directory from the parallel THAPBI repository.

As the Python framework for adapting that script will be used for
development of a more flexible tool, we are breaking it out into
its own repository. Several files will be modified and renamed as
development progresses.