add tests for blast.py module
This commit adds rudimentary tests for the Blastclust class, which
required the creation of a new NotExecutableError, and modifications
to the way that logging works with this class (I'd like to see this
extend to the other wrapper classes). Module components shouldn't
require a logger object for operation, so Blastclust() now raises
an error, rather than sending a logger message and using sys.exit(1).
Capturing that error gracefully should happen in the calling script.
Changes were made in run-pycits.py to turn the logger off for this
wrapper.