fix test_metapy_tool_db_len_ass_len.py
This fix required reworking function interfaces in metapy_tools.py
as well as changes to how the tests were performed.

NOTE: when we need to check float values, it is bad practice to
convert them to strings, then check that against a string. Always
use an approximate equality test function (e.g. assert_almost_equals())
with a stated tolerance!